Fix translation of structs: since they are represented as references, no need to box them. So deleted BoxFromStruct and Unbox2Struct.
Simplified the body of the generated copy_ctor for structs by just assuming
taht the dynamic type of the copy is the same as the dynamic type of the
argument.

Make sure that the copy_ctor for a struct is defined only for the template
and not for any generic instantiations.