Fix file path issues in the regression test under Visual Studio 2015.
The test was looking for RegressionTestInput.dll in the test output
directory without having copied it there, and it was writing the actual
outputs to TranslationTest\bin\Debug instead of the test output directory.
I'm not sure how this was supposed to work before, but hopefully this
change does not break any other use modes.